Extract LeafletCustomMarker component so it can be used by Hurricane marker
[#163254310]
[#163237654]
[#163237593]
