+++ src/components/app.scss
@@ -1,3 +1,4 @@
 .app {
   height: 100%;
+  font-family: Helvetica, Arial, sans-serif;
 }
